requirements.txt
# Python dependencies
pandas
numpy
scikit-learn
matplotlib
seaborn
pyyaml
joblib
pytest
fastapi
uvicorn
